+++ src/background.js
@@ -19,6 +19,8 @@ const onBootstrapEnd = () => {
 const timeoutBootstrap = debounce(onBootstrapEnd, 5000)
 
 chrome.runtime.onMessage.addListener(function(request, sender, sendResponse) {
+  console.log(request)
+
   if (request) {
     switch (request.action) {
       case 'bootstrap-start':
@@ -44,6 +46,7 @@ chrome.runtime.onMessage.addListener(function(request, sender, sendResponse) {
       case 'yt-popout':
         createYoutubePopout(request.payload)
         break
+      default:
     }
   }
 })