+++ src/components/storage-management-card.vue
@@ -2,10 +2,10 @@
   <b-card title="Storage management">
     <form>
       <b-form-group label="Delete video archive after X  days">
-        <b-input type="number" placeholder="days" v-model="storageLifeTime"> </b-input>
+        <b-input type="number" placeholder="days" v-model="storageLifetime"> </b-input>
       </b-form-group>
       <b-form-group label="Limit to X message per video.">
-        <b-input type="number" placeholder="Message limit" v-model="limitMsgPerVideo"> </b-input>
+        <b-input type="number" placeholder="Message limit" v-model="limitMsgPerVideo" min="0" step="1"> </b-input>
       </b-form-group>
       <b-alert variant="warning" show>
         You can put "0" for unlimited, but take care of not hitting the storage limit, else the extension will no longer work.
@@ -21,15 +21,21 @@ export default {
         return this.$store.getters.global.limitMsgPerVideo
       },
       set(value) {
+        if (!value) {
+          value = 0
+        }
         this.$store.commit('updateGlobal', { value, path: 'limitMsgPerVideo' })
       },
     },
-    storageLifeTime: {
+    storageLifetime: {
       get() {
-        return this.$store.getters.global.storageLifeTime
+        return this.$store.getters.global.storageLifetime
       },
       set(value) {
-        this.$store.commit('updateGlobal', { value, path: 'storageLifeTime' })
+        if (!value) {
+          value = 0
+        }
+        this.$store.commit('updateGlobal', { value, path: 'storageLifetime' })
       },
     },
   },