+++ webpack.config.js
@@ -14,6 +14,7 @@ const config = {
     background: './background.js',
     content: './content.js',
     loading: './loading.js',
+    inject: './inject.js',
     'popout-main': './popout-main.js',
     'content-style': './content-style.scss',
     popout: './popout.scss',